Commit Raspberry Pi Java app for controlling the Arduino
This RPi Java app is used to control the Arduino running the LedTable
code. It connects via usb for a serial connection like when using the
Arduino IDE.

Pi4J is required and the jars should be included in lib/pi4j

Some test animations are included. The app includes a feature where it
can read bmp images for display, or a folder of bmp images to loop
through to simulate animations.