Rewrite to use webservice, remove sql
Rewrite of a lot of the code, completely removing any sql dependencies.
The polling was a not a great idea, it worked but added too much
unnecessary complexity.

Added an ant build script which can build the jar in eclipse. This is a
lot faster than the run script from before, since it can be ran on a
different pc and then the jar uploaded to the pi.

I started working on a test gui to help with debugging the table output.

I started working on a midi display mode. This is really early and
doesn't work well.